% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calculate_clusters.R
\name{calculate_clusters}
\alias{calculate_clusters}
\title{Function to cluster data}
\usage{
calculate_clusters(input, output, session, data, vars, cluster_type)
}
\arguments{
\item{data}{data frame which to use for clustering}

\item{vars}{variables to use for clustering}

\item{cluster_type}{which clustering algorithm to use, multiple allowed}

\item{input/output/session}{standard shiny parameters to allow interactivity}
}
\value{
data frame with each column being a different clustering method
}
\description{
Function to cluster data
}
