% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calculate_dm.R
\name{calculate_dm}
\alias{calculate_dm}
\title{Function to calculate Diffusion Map}
\usage{
calculate_dm(input, output, session, data, vars, dims, cluster_df)
}
\description{
Function to calculate Diffusion Map
}
