% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calculate_heatmap.R
\name{calculate_heatmap}
\alias{calculate_heatmap}
\title{Function to calculate and render heatmap}
\usage{
calculate_heatmap(input, output, session, data, vars, show_clus)
}
\description{
Function to calculate and render heatmap
}
