% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calculate_umap.R
\name{calculate_umap}
\alias{calculate_umap}
\title{Function to calculate UMAP}
\usage{
calculate_umap(input, output, session, data, vars, dims, cluster_df)
}
\description{
Function to calculate UMAP
}
