% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_edit_colnames.R
\name{edit_colnames}
\alias{edit_colnames}
\title{Function to edit data frame column names}
\usage{
edit_colnames(data = NULL, fs = NULL, input, output, session)
}
\arguments{
\item{data}{data frame}
}
\value{
a modal dialogue to edit colnames
}
\description{
Function to edit data frame column names
}
