% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_edit_samplenames.R
\name{edit_samplenames}
\alias{edit_samplenames}
\title{Funtion to display popup modal allowing user to edit sample names}
\usage{
edit_samplenames(data = NULL, input, output, session, ...)
}
\arguments{
\item{data}{data frame}
}
\value{

}
\description{
Funtion to display popup modal allowing user to edit sample names
}
