% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_fs_to_ff.R
\name{fs_to_ff}
\alias{fs_to_ff}
\title{Function to convert flowSet to flowFrame}
\usage{
fs_to_ff(fs = NULL)
}
\arguments{
\item{flowSet}{imput flowSet object}
}
\value{
flowFrame
}
\description{
Function to convert flowSet to flowFrame
}
