% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_ui.R
\name{golem_add_external_resources}
\alias{golem_add_external_resources}
\title{Add external Resources to the Application}
\usage{
golem_add_external_resources()
}
\description{
Add external Resources to the Application
}
