% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R, R/app_ui.R
\docType{data}
\name{packages_1}
\alias{packages_1}
\title{Application Server}
\format{
An object of class \code{character} of length 24.

An object of class \code{character} of length 24.
}
\usage{
packages_1

packages_1
}
\description{
Application Server

Application User Interface
}
\keyword{datasets}
