% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\docType{data}
\name{rv}
\alias{rv}
\title{Utility Functions
Store reactive values in rv list}
\format{
An object of class \code{reactivevalues} of length 3.
}
\usage{
rv
}
\description{
Utility Functions
Store reactive values in rv list
}
\keyword{datasets}
