% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_transform_data.R
\name{transform_data}
\alias{transform_data}
\title{Function to transform data}
\usage{
transform_data(ff = NULL, method = NULL)
}
\arguments{
\item{ff}{Input flowframe to be transformed}

\item{method}{transformation method}
}
\value{
transformed flowframe
}
\description{
Function to transform data
}
